pyproject.toml
[project]
name = "audio"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "librosa>=0.11.0",
    "lightning>=2.4.0",
    "numpy>=2.3.5",
    "pandas>=3.0.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "transformers>=5.0.0",
]
